refactor: component names for process definition view
